Scripts/pyproject.toml
[project]
name = "so-arm101-llm-integration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohappyeyeballs>=2.6.1",
    "aiohttp>=3.12.15",
    "aiosignal>=1.4.0",
    "annotated-types>=0.7.0",
    "attrs>=25.3.0",
    "av>=15.1.0",
    "blinker>=1.9.0",
    "certifi>=2025.8.3",
    "charset-normalizer>=3.4.3",
    "click>=8.3.0",
    "cloudpickle>=3.1.1",
    "cmake>=4.1.0",
    "datasets>=3.6.0",
    "deepdiff>=8.6.1",
    "diffusers>=0.35.1",
    "dill>=0.3.8",
    "draccus>=0.10.0",
    "einops>=0.8.1",
    "farama-notifications>=0.0.4",
    "feetech-servo-sdk>=1.0.0",
    "ffmpeg>=1.4",
    "filelock>=3.19.1",
    "flask>=3.1.2",
    "frozenlist>=1.7.0",
    "fsspec>=2025.3.0",
    "gitdb>=4.0.12",
    "gitpython>=3.1.45",
    "gymnasium>=0.29.1",
    "hf-transfer>=0.1.9",
    "hf-xet>=1.1.10",
    "huggingface-hub>=0.35.0",
    "idna>=3.10",
    "imageio>=2.37.0",
    "imageio-ffmpeg>=0.6.0",
    "importlib-metadata>=8.7.0",
    "inquirerpy>=0.3.4",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "jsonlines>=4.0.0",
    "lerobot>=0.3.3",
    "markupsafe>=3.0.2",
    "mergedeep>=1.3.4",
    "mpmath>=1.3.0",
    "multidict>=6.6.4",
    "multiprocess>=0.70.16",
    "mypy-extensions>=1.1.0",
    "networkx>=3.5",
    "numpy>=2.3.3",
    "opencv-python-headless>=4.11.0.86",
    "orderly-set>=5.5.0",
    "packaging>=25.0",
    "pandas>=2.3.2",
    "pfzy>=0.3.4",
    "pillow>=11.3.0",
    "platformdirs>=4.4.0",
    "prompt-toolkit>=3.0.52",
    "propcache>=0.3.2",
    "protobuf>=6.32.1",
    "psutil>=7.1.0",
    "pyarrow>=21.0.0",
    "pydantic>=2.11.9",
    "pydantic-core>=2.33.2",
    "pynput>=1.8.1",
    "pyobjc-core>=11.1",
    "pyobjc-framework-applicationservices>=11.1",
    "pyobjc-framework-cocoa>=11.1",
    "pyobjc-framework-coretext>=11.1",
    "pyobjc-framework-quartz>=11.1",
    "pyserial>=3.5",
    "python-dateutil>=2.9.0.post0",
    "pytz>=2025.2",
    "pyyaml>=6.0.2",
    "pyyaml-include>=1.4.1",
    "regex>=2025.9.18",
    "requests>=2.32.5",
    "rerun-sdk>=0.22.1",
    "safetensors>=0.6.2",
    "sentry-sdk>=2.38.0",
    "setuptools>=80.9.0",
    "six>=1.17.0",
    "smmap>=5.0.2",
    "sympy>=1.14.0",
    "termcolor>=3.1.0",
    "toml>=0.10.2",
    "torch>=2.7.1",
    "torchcodec>=0.5",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
    "typing-extensions>=4.15.0",
    "typing-inspect>=0.9.0",
    "typing-inspection>=0.4.1",
    "tzdata>=2025.2",
    "urllib3>=2.5.0",
    "uv>=0.8.18",
    "wandb>=0.22.0",
    "wcwidth>=0.2.13",
    "werkzeug>=3.1.3",
    "xxhash>=3.5.0",
    "yarl>=1.20.1",
    "zipp>=3.23.0",
]
